Safari-611.1.10/LayoutTests/http/tests/cache/disk-cache/disk-cache-media.html

<script src="/js-test-resources/js-test-pre.js"></script>
<script src="resources/cache-test.js"></script>
<body>
<script>

var testMatrix =
[
 [
  { responseHeaders: {'Cache-control': 'max-age=0' } },
  { responseHeaders: {'Cache-control': 'max-age=100' } },
  ],
 [
  { responseHeaders: {'Content-Type': 'text/plain' } },
  { responseHeaders: {'Content-Type': 'video/mp4' } },
  { responseHeaders: {'Content-Type': 'audio/mp4' } },
  ],
 ];

description("Tests that media resources loaded via XHR is not cached.");

var tests = generateTests(testMatrix);

debug("running " + tests.length + " tests");
debug("");

runTests(tests);

</script>
<script src="/js-test-resources/js-test-post.js"></script>
